pyproject.toml
[project]
name = "salary-project"
version = "0.1.0"
description = ""
authors = [
    {name = "N1ntek",email = "zobalarm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncmy (>=0.2.10)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
]

[tool.poetry]
package-mode=false

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.11.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
